added and configured hibernate  - create PrepareDataApp  - create Dao  - now the list of products is generated from the database  - now products are added to the database
