+++ src/main/java/ru/geekbrains/webapp/repository/ProductRepository.java
@@ -1,73 +1,37 @@
 package ru.geekbrains.webapp.repository;
 
-import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.hibernate.cfg.Configuration;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import ru.geekbrains.webapp.model.Product;
+import ru.geekbrains.webapp.repository.preparation.PrepareDataApp;
 
 import javax.annotation.PostConstruct;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Repository
 public class ProductRepository {
-
-    private SessionFactory factory;
     private List<Product> catalog;
+    private ProductDAO dao;
 
-    @Autowired
-    public ProductRepository() {
-        this.factory = new Configuration()
-                .configure("config/hibernate.xml")
-                .buildSessionFactory();
-    }
 
     @PostConstruct
     public void init() {
-        Session session;
-
-        try {
-            String sql = Files.lines(Paths.get("CreateTableProducts.sql")).collect(Collectors.joining(" "));
-            session = factory.getCurrentSession();
-            session.beginTransaction();
-            session.createNativeQuery(sql).executeUpdate();
-            session.getTransaction().commit();
-            catalog = findProduct();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        catalog = dao.findProduct();
     }
 
     public List<Product> findAll() {
         return Collections.unmodifiableList(catalog);
     }
 
-    public List<Product> findProduct() {
-        List<Product> list;
-        try (Session session = factory.getCurrentSession()) {
-            session.beginTransaction();
-            list = session.createQuery("select a from Product a", Product.class).getResultList();
-            session.getTransaction().commit();
-        }
-        return list;
-    }
 
     public Product findProductById(Long id) {
         return catalog.stream().filter(product -> product.getId() == id).findFirst().get();
     }
 
     public void addProduct(Product product) {
-        try (Session session = factory.getCurrentSession()){
-            session.beginTransaction();
-            session.save(product);
-            session.getTransaction().commit();
-        }
-        catalog = findProduct();
+        dao.addProduct(product);
+        catalog = dao.findProduct();
     }
 
     public void addCost(int id) {
@@ -84,5 +48,12 @@ public class ProductRepository {
         catalog.get(id).setCost(beginValue);
     }
 
+    public ProductDAO getDao() {
+        return dao;
+    }
 
+    @Autowired
+    public void setDao(ProductDAO dao) {
+        this.dao = dao;
+    }
 }
